+++ CHANGES.rst
@@ -4,7 +4,13 @@ Changelog
 1.3.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+New:
+
+- *add item here*
+
+Fixes:
+
+- *add item here*
 
 
 1.3.3 (2015-09-09)