+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 New:
 
-- *add item here*
+- Add compatibility with Python 2 and Python 3.
+  [gforcada]
 
 Fixes: